requirements.txt
python-telegram-bot>=20.0
flask
psycopg2-binary>=2.9.3